third_party/blink/web_tests/fast/files/file-reader-detached-no-crash.html

<!DOCTYPE HTML>
<body>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<script>
description("Accessing FileReader when in a detached state.");

window.jsTestIsAsync = true;

var w;
var fileReader;
function processMessage(event) {
    if (event.data == "opened") {
        fileReader = new w.FileReader();
        w.close();
    } else if (event.data == "closed") {
        shouldThrow("fileReader.readAsArrayBuffer(new File(['hello'], 'world.html'))");
        finishJSTest();
    }
}

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.setPopupBlockingEnabled(false);
}
w = window.open('../../resources/window-postmessage-open-close.html');
window.addEventListener("message", processMessage, false);
</script>
</body>
